rpc: api: Create flag file on LC during reboot via supervisor
When reboot command is issued on supervisor, through RPC calls,
supervisor issues reboot command on LCs. show reboot-cause currently
doesn't capture this information and shows an empty string for user
and not enough information for cause of reboot.

Through this patch, we create a flag file, which would then be
checked by LC on startup during 'determine-reboot-cause' and if
the file is present, it gleans that supervisor has issued previous
reboot command, updates the reboot-cause and deletes the flag
file.

Change-Id: I2225a9a46959ebe24e891744df9867bf50f26477
Signed-off-by: Mohan Yelugoti <ymd@arista.com>