Modify LC reboot cause on ungraceful supervisor reboot
When supervisor boots up, we create a flag file and populate it with
reboot cause for LCs. If supervisor reboots because there was a kernel
panic, reboot cause for LCs is filled with Heartbeat loss in the flag
file, in all other cases we fill it with None.

When LC comes up, it interacts with supervisor using Rpc and reads the
flag file for its reboot cause, and writes this to reboot-cause.txt,
which is later processed in determine-reboot-cause.

Change-Id: Iee5b837bd2eb7a44cd99a815cab097347e008e7a
Signed-off-by: Mohan Yelugoti <ymd@arista.com>