cooling: improve cooling algorithm stability
Fix an issue to ensure update runs every iteration
Currently the update only happens when we hit the garbage collection
threshold which was not intended.
The mitigation is to set gc_count to 1

Update and make configurable the temperature delta for which the
algorithm start adjusting the fans.
Prior the algo would stop adjusting fan speed when the delta from target
to overheat was 5%. It has now changed to 5% below target and 0% above.
These are now configurable as well.

Change-Id: Id1dd7cebe7157ee716aa7335be57b3a54a371b84