Improve LC reboot cause when chassis is powercycled
When chassis is powercycled, and when supervisor it comes up, it
powercycles any LC which are still on. Currently, LC shows the
rebootCause as Thermal overload, which is misleading.

In this patch, during platform-reboot-cause.service, we also
provide hardware report to calls which handle ungraceful reboot,
so we would also have HW reboot cause while making the decision.

This patch aims to extend current mechanism of displaying
'Heartbeat lost with supervisor' on LC when supervisor reboots
due to kernel panic (ungraceful reboot) to also display the
same message when Chassis is powercycled.

Change-Id: I31efdccab9bcee9ae328a248e6ba733900b845be
Signed-off-by: Mohan Yelugoti <ymd@arista.com>