cooling: change data export mechanism
Improve the mechanism that exports cooling data for plotting.
The new mechanism adds lines to a CSV every iteration of the cooling
algorithm.
The old mechanism was keeping all the data in memory and then dumping
everything in a json file every iteration.
This is a big improvement both in terms of memory usage and
computational resources.

Change-Id: I31176c40234243605cf70abf230e8f909f910706