sensors: add optional rate-limiting to sensors
The SensorDesc class takes an optional readFilter parameter that will
be used to apply a filter to readings. One filter is provided for now:
it prevents sensor values changing by too much between readings; it
can be used to prevent thermalctld from complaining that the CPU
temperature changes too much. The Crow CPU is updated to apply this
filter, as an example.

The JSON encoder for the `arista dump` command is changed to ignore any
non-basic Python types so that it doesn't fail when it encounters a
filter object.

Change-Id: I8cb075115eccf95ac9bc3a72e2f210cd2a54fc9b