leds: use Linux LED device classes as intended
Instead of overloading the `brightness` attribute of LED devices to be
used for changing colours, use the kernel's support for multicolour
LEDs. Also refactor mono LEDs to use the attributes as designed.

Userspace now needs to tell the SCD driver what kind of LED is present
when it creates the LED objects. Most of the defined kinds are legacy
and it is expected that most future platforms will use 8-bit RGB LEDs
with the same programming interface.

For now, only Moby is changed to use the new LED support.

Change-Id: I090d3312947748d1c7ff97c8fbefb4f2da0f9a82