quirks: refactor quirk subsystem
Make the quirk facility available for all components.
The `quirks` attribute should now be always available.
Refactor the Quirk base classes to reduce the amout of code needed in
child definitions.

Allow the quirk to execute at 4 different time.
Before or after the driver initialization, after the device is taken out
of reset, or later once the system is running.

Introduce a new test that validates the quirks for all platforms.

Change-Id: I478b4e4e0bff8def080e239daff8a7a324f122fc